added nice icon on button

+++ CookBook/ChiefListOfDishesForm.cs
@@ -44,15 +44,30 @@ namespace CookBook {
             // кнопка удаления
             foreach (DishCard card in this.LayoutPanel.Controls.OfType<DishCard>()) {
                 Button removeBtn = new Button();
-                removeBtn.Text = "D";
+                //removeBtn.Text = "D";
+                removeBtn.BackgroundImage = Properties.Resources._8664938_trash_can_delete_remove_icon;
+                removeBtn.BackgroundImageLayout = ImageLayout.Center;
+                //removeBtn.FlatStyle = FlatStyle.Flat;
+
                 removeBtn.Click += (object s, EventArgs ev) => {
-                    DBManager.DeleteRecord(card.ID);
-                    this.LayoutPanel.Controls.Remove(card);
+                    DialogResult result = MessageBox.Show($"Вы уверены что хотите удалить {card._dishFields.Name}",
+                        "",
+                        MessageBoxButtons.YesNo,
+                        MessageBoxIcon.Information,
+                        MessageBoxDefaultButton.Button2
+                        ); //
+                    if (result == DialogResult.Yes) {
+                        DBManager.DeleteRecord(card.ID);
+                        this.LayoutPanel.Controls.Remove(card);
+                    }
                 };
                 card.AddBtn(removeBtn);
 
                 Button changeBtn = new Button();
-                changeBtn.Text = "C";
+                //changeBtn.Text = "C";
+                changeBtn.BackgroundImage = Properties.Resources._2931178_change_edit_pencil_creative_design_icon__1_;
+                changeBtn.BackgroundImageLayout = ImageLayout.Center;
+
                 changeBtn.Click += (object s, EventArgs ev) => {
                     //this.Hide();
                     AddDishForm addDishForm = new AddDishForm(card._dishFields);
